// src/llm/types/xiaomi-model/response.ts
export type XOAIChatCompletionToolCallDelta = XOAIAssistantToolCall & {
  /**
   * The index of the called tool in the tool_calls list, starting from 0.
   */
  index: number;
};
// ...
export type BaseXOAIChatCompletionChoice = {
  /**
   * The reason the model stopped generating tokens.
   *
   * This will be:
   *
   * **stop** if the model hit a stop point or a provided stop sequence
   *
   * **length** if the maximum number of tokens specified in the request was reached
   *
   * **tool_calls** if the model called a tool
   *
   * **content_filter** if content was omitted due to a flag from our content filters
   *
   * **repetition_truncation** if the model detects repetition.
   */
  finish_reason?: string | null;
  /**
   * The index of the choice in the list of choices.
   */
  index: number;
};

export type XOAIChatCompletionChoice = {
  /**
   * A chat completion message generated by the model.
   */
  message: XOAIChatCompletionMessage;
} & BaseXOAIChatCompletionChoice;

export type XOAIChatCompletionUsagePromptTokensDetails = {
  /**
   * Number of tokens served from cache.
   */
  cached_tokens?: number;
  /**
   * Audio input tokens present in the prompt.
   */
  audio_tokens?: number;
  /**
   * Image input tokens present in the prompt.
   */
  image_tokens?: number;
  /**
   * Video input tokens present in the prompt.
   */
  video_tokens?: number;
};

export type XOAIChatCompletionUsageCompletionTokensDetails = {
  /**
   * Tokens generated by the model for reasoning.
   */
  reasoning_tokens?: number;
};

export type XOAIChatCompletionUsage = {
  /**
   * Number of tokens in the generated completion.
   */
  completion_tokens: number;
  /**
   * Number of tokens in the prompt.
   */
  prompt_tokens: number;
  /**
   * Total number of tokens used in the request (prompt + completion).
   */
  total_tokens: number;
  /**
   * Breakdown of tokens used in a completion.
   */
  completion_tokens_details?: XOAIChatCompletionUsageCompletionTokensDetails;
  /**
   * Breakdown of tokens used in the prompt.
   */
  prompt_tokens_details?: XOAIChatCompletionUsagePromptTokensDetails | null;
  /**
   * Detailed usage of the web search API.
   */
  web_search_usage?: {
    /**
     * Number of API calls in web search.
     */
    tool_usage?: number;
    /**
     * Number of web pages returned by the web search API.
     */
    page_usage?: number;
  };
};
// ...
export type XOAIChatCompletionDelta = {
  /**
   * The contents of the chunk message.
   */
  content?: string;
  /**
   * The reasoning contents of the assistant message, before the final answer.
   */
  reasoning?: string;
  /**
   * The reasoning contents of the assistant message, before the final answer.
   */
  reasoning_content?: string;
  /**
   * The role of the author of this message.
   */
  role?: 'assistant';
  /**
   * The tools to be called by the model and the parameters required for the calls.
   * It can contain one or more tool response objects.
   */
  tool_calls?: XOAIChatCompletionToolCallDelta[];
  /**
   * If the audio output modality is requested, this object contains data about the audio response from the model.
   */
  audio?: Partial<XOAIChatCompletionAudio> | null;
  /**
   * The final audio broadcast text after intelligent optimization and polishing.
   * This field is only returned when the request parameter optimize_text_preview is set to true.
   */
  final_text_preview?: string;
  /**
   * After web search, the model returns annotations for all referenced URLs.
   */
  annotations?: XOAIChatCompletionAnnotation[];
  /**
   * Error message of web search.
   */
  error_message?: string;
};

export type XOAIChatCompletionChunkChoice = {
  /**
   * A chat completion delta generated by streamed model responses.
   */
  delta: XOAIChatCompletionDelta;
} & BaseXOAIChatCompletionChoice;

export type XOAIChatCompletionChunk = {
  /**
   * A unique identifier for the chat completion. Each chunk has the same ID.
   */
  id: string;
  /**
   * The object type, which is always chat.completion.chunk.
   */
  object: 'chat.completion.chunk';
  /**
   * The Unix timestamp, in seconds, of when the chat completion was created. Each chunk has the same timestamp.
   */
  created: number;
  /**
   * The model to generate the completion.
   */
  model: string;
  /**
   * A list of chat completion choices.
   */
  choices: XOAIChatCompletionChunkChoice[];
  /**
   * Usage statistics for the completion request.
   */
  usage?: XOAIChatCompletionUsage | null;
};